test: ProductField jest unit tests
